% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{format_channel_reads_plot}
\alias{format_channel_reads_plot}
\title{Function to combine and format channel reads and reference data ready for plotting}
\usage{
format_channel_reads_plot(channel_reads, phase_groups, device_hierarchy)
}
\arguments{
\item{channel_reads}{Channel reads dataframe}

\item{phase_groups}{Phase groups dataframe}

\item{device_hierarchy}{Device hierarchy dataframe}
}
\value{
dataframe
}
\description{
Function to combine and format channel reads and reference data ready for plotting
}
