% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{format_event_data}
\alias{format_event_data}
\title{Function to format event data for presentation / csv export}
\usage{
format_event_data(event_data, device_hierarchy)
}
\arguments{
\item{event_data}{Dataframe from get_event_data}

\item{device_hierarchy}{Dataframe from get_device_hierarchy}
}
\value{
dataframe
}
\description{
Function to format event data for presentation / csv export
}
\examples{
format_event_data(get_event_data(meter = '00135003001e913d', event_type = c('15565','15566','15567','15568'), start_time = as.POSIXct('2018-05-01 00:00:00', tz = 'Australia/Melbourne'), end_time = as.POSIXct('2018-05-25 23:59:00', tz = 'Australia/Melbourne')),get_device_hierarchy(parse_device_string('00135003001e913d')))
}
