% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{format_voltage_histogram_data}
\alias{format_voltage_histogram_data}
\title{Function to format histogram data for plots}
\usage{
format_voltage_histogram_data(histogram_data, voltage_offset = 0)
}
\arguments{
\item{histogram_data}{Histogram dataframe}

\item{voltage_offset}{Integer value to shift voltages}
}
\value{
Dataframe
}
\description{
Function to format histogram data for plots
}
\examples{
format_voltage_histogram_data(get_zone_substation_voltage_histogram_data(zone_substation = 'CDA', interval_ts = as.POSIXct('2018-05-01 12:30:00')), voltage_offset = 5)
}
