% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{get_channel_amps_plot_plotly}
\alias{get_channel_amps_plot_plotly}
\title{Function to get amps plotly plot}
\usage{
get_channel_amps_plot_plotly(cd, aggregate_by_phase = F)
}
\arguments{
\item{cd}{output dataframe from format_channel_reads_plot()}
}
\value{
plotly object
}
\description{
Function to get amps plotly plot
}
