% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{get_channel_imp_plot_gg}
\alias{get_channel_imp_plot_gg}
\title{Function to get impedance ggplot plot}
\usage{
get_channel_imp_plot_gg(cd)
}
\arguments{
\item{cd}{output dataframe from format_channel_reads_plot()}
}
\value{
ggplot object
}
\description{
Function to get impedance ggplot plot
}
