% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_channel_reads}
\alias{get_channel_reads}
\title{Function to get channel reads}
\usage{
get_channel_reads(channels, start_time, end_time)
}
\arguments{
\item{channels}{A vector of nic channel ids to pull channel reads for}

\item{start_time}{Start of period to pull channel reads for}

\item{end_time}{End of period to pull channel reads for}
}
\value{
dataframe
}
\description{
Function to get channel reads
}
\examples{
get_channel_reads(channels = c('n0013500300135a31_a','0013500300101ce4_b'), start_time = as.POSIXct('2018-01-01 12:30:00', tz='Australia/Melbourne'),end_time = as.POSIXct('2018-01-01 13:30:00', tz='Australia/Melbourne'))
}
