% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{get_channel_waveform_parameters}
\alias{get_channel_waveform_parameters}
\title{Function to get channel waveform ggplot parameters, this may take environment parameters in future}
\usage{
get_channel_waveform_parameters()
}
\value{
list
}
\description{
Function to get channel waveform ggplot parameters, this may take environment parameters in future
}
