% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_device_hierarchy}
\alias{get_device_hierarchy}
\title{Function to get channel hierarchy information for a list of channels, meters, serialnums, nmis}
\usage{
get_device_hierarchy(devices)
}
\arguments{
\item{devices}{A list containing channels, meters,serial_nums and nmis}
}
\value{
dataframe
}
\description{
Function to get channel hierarchy information for a list of channels, meters, serialnums, nmis
}
\examples{
get_device_hierarchy(parse_device_string("0013500300101ce4_b\\n0013500300135a31_a\\n00135003001144df\\n01254662\\n6407275764"))
get_device_hierarchy(parse_device_string("0013500300101ce4_b"))
}
