% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_device_waveform_export}
\alias{get_device_waveform_export}
\title{Function to get device waveform data export}
\usage{
get_device_waveform_export(channel_reads, phase_groups, device_hierarchy)
}
\arguments{
\item{channel_reads}{Channel reads dataframe}

\item{phase_groups}{Phase groups dataframe}

\item{device_hierarchy}{Device hierarchy dataframe}
}
\value{
dataframe
}
\description{
Function to get device waveform data export
}
