% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{get_device_waveform_plotly}
\alias{get_device_waveform_plotly}
\title{Function to get device plotly plot}
\usage{
get_device_waveform_plotly(channel_reads, phase_groups, device_hierarchy,
  aggregate_amps = F, include_imp = F, include_pf = F)
}
\arguments{
\item{channel_reads}{Channel reads dataframe}

\item{phase_groups}{Phase groups dataframe}

\item{device_hierarchy}{Device hierarchy dataframe}
}
\value{
plotly object
}
\description{
Function to get device plotly plot
}
