% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_distribution_substation}
\alias{get_distribution_substation}
\title{Function to get distribution substations}
\usage{
get_distribution_substation()
}
\value{
dataframe
}
\description{
Function to get distribution substations
}
\examples{
get_distribution_substation()
}
