% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_distribution_substation_channels}
\alias{get_distribution_substation_channels}
\title{Function to get channels under a distribution substation}
\usage{
get_distribution_substation_channels(distribution_substation)
}
\arguments{
\item{distribution_substation}{The distribution substation name}
}
\value{
dataframe
}
\description{
Function to get channels under a distribution substation
}
\examples{
get_distribution_substation_channels('MILNE-JAMES')
}
