% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_distribution_substation_voltage_histogram_data}
\alias{get_distribution_substation_voltage_histogram_data}
\title{Function to get distribution substation histogram data}
\usage{
get_distribution_substation_voltage_histogram_data(distribution_substation,
  interval_ts)
}
\arguments{
\item{distribution_substation}{The distribution substation name}

\item{interval_ts}{Timestamp to get zone substation histogram}
}
\value{
Dataframe
}
\description{
Function to get distribution substation histogram data
}
\examples{
get_distribution_substation_voltage_histogram_data(distribution_substation = 'MILNE-JAMES', interval_ts = as.POSIXct('2018-05-01 12:30:00'))
}
