% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_distribution_transformer_lv_circuit}
\alias{get_distribution_transformer_lv_circuit}
\title{Function to get transformers under a distribution substation}
\usage{
get_distribution_transformer_lv_circuit(distribution_transformer_id)
}
\arguments{
\item{distribution_transformer_id}{The distribution transformer id}
}
\value{
dataframe
}
\description{
Function to get transformers under a distribution substation
}
\examples{
get_distribution_transformer_cuit('000000000004014234')
}
