% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_event_data}
\alias{get_event_data}
\title{Function to get event data for supplied meters for a given period of time}
\usage{
get_event_data(meter = c(), event_type = c(), start_time, end_time)
}
\arguments{
\item{meter}{A vector of nic mac ids}

\item{event_type}{A vector of event type ids}

\item{start_time}{Datetime start time of period}

\item{end_time}{Datetime end time of period}
}
\value{
dataframe
}
\description{
Function to get event data for supplied meters for a given period of time
}
\examples{
get_event_data(meter = c('00135003001e913d'), event_type = c('15565','15566','15567','15568'), start_time = as.POSIXct('2018-05-01 00:00:00', tz = 'Australia/Melbourne'), end_time = as.POSIXct('2018-05-25 23:59:00', tz = 'Australia/Melbourne'))
}
