% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_lv_circuit_channels}
\alias{get_lv_circuit_channels}
\title{Function to get channels under lv circuits}
\usage{
get_lv_circuit_channels(lv_circuit)
}
\value{
dataframe
}
\description{
Function to get channels under lv circuits
}
\examples{
get_lv_circuit_channels(c("112807948","140987703"))
}
