% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_meter_details_ws}
\alias{get_meter_details_ws}
\title{Function to get meter hierachy information as requested by webservice}
\usage{
get_meter_details_ws(input_string)
}
\arguments{
\item{input_string}{A string containing comma separated channels, nic mac ids, serial numbers and nmis}
}
\value{
dataframe
}
\description{
Function to get meter hierachy information as requested by webservice
}
\examples{
get_meter_details_ws('01332167')
}
