% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{get_meter_waveform_plot_gg}
\alias{get_meter_waveform_plot_gg}
\title{Function to get meter waveform ggplot}
\usage{
get_meter_waveform_plot_gg(channel_reads, device_hierarchy,
  filter_voltage = F, aggregate_amps = F, include_imp = F,
  include_pf = F)
}
\arguments{
\item{channel_reads}{Channel reads dataframe}

\item{device_hierarchy}{Device hierarchy dataframe}

\item{filter_voltage}{Boolean flag on whether voltages should be filtered}

\item{aggregate_amps}{Boolean flag on whether amps should be aggregated}

\item{include_imp}{Boolean flag on whether the plot should include an impedance plot}

\item{include_pf}{Boolean flag on whether the plot should include a power factor plot}
}
\value{
cowplot object
}
\description{
Function to get meter waveform ggplot
}
