% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_nap_api_url}
\alias{get_nap_api_url}
\title{Function to get the nap api url configured in the system environment}
\usage{
get_nap_api_url()
}
\value{
character
}
\description{
Function to get the nap api url configured in the system environment
}
\examples{
get_nap_api_url()
}
