% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_node_channels}
\alias{get_node_channels}
\title{Function to get channels under a topology id}
\usage{
get_node_channels(topology_id)
}
\arguments{
\item{topology_id}{Integer of the topology node id (gis)}
}
\value{
dataframe
}
\description{
Function to get channels under a topology id
}
\examples{
get_node_channels(86223764)
}
