% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_phase_colours}
\alias{get_phase_colours}
\title{Function to get the phase colours configured in the system environment}
\usage{
get_phase_colours()
}
\value{
named vector
}
\description{
Function to get the phase colours configured in the system environment
}
\examples{
get_phase_colours()
}
