% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_phase_groups}
\alias{get_phase_groups}
\title{Function to calculate phase group clusters}
\usage{
get_phase_groups(channel_reads)
}
\arguments{
\item{channel_reads}{Channel reads dataframe to calculate phase groups for}
}
\value{
dataframe
}
\description{
Function to calculate phase group clusters
}
