% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_pole}
\alias{get_pole}
\title{Function to get pole information}
\usage{
get_pole(pole_id)
}
\arguments{
\item{pole_id}{The pole id (lis number)}
}
\value{
dataframe
}
\description{
Function to get pole information
}
\examples{
get_pole(1109226)
}
