% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{get_substation_waveform_plotly}
\alias{get_substation_waveform_plotly}
\title{Function to get substation waveform plotly plot}
\usage{
get_substation_waveform_plotly(channel_reads, phase_groups, device_hierarchy,
  aggregate_amps = T, include_imp = F, include_pf = F,
  filter_voltage = T)
}
\arguments{
\item{channel_reads}{Channel reads dataframe}

\item{phase_groups}{Phase groups dataframe}

\item{device_hierarchy}{Device hierarchy dataframe}
}
\value{
plotly object
}
\description{
Function to get substation waveform plotly plot
}
