% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{get_voltage_histogram_plot_gg}
\alias{get_voltage_histogram_plot_gg}
\title{Function to get a histogram plot}
\usage{
get_voltage_histogram_plot_gg(histogram_data, voltage_offset = 0,
  reference_lines = c(), chart_title = "", percentiles = c(1, 99))
}
\arguments{
\item{histogram_data}{Voltage histogram dataframe}

\item{voltage_offset}{Integer value to shift voltages}

\item{reference_lines}{Vector of voltages to draw horizontal lines for on the plot}

\item{chart_title}{Chart title text to show above plot}

\item{percentiles}{Vector of percentiles to calculate and include in plot subtitle}
}
\value{
ggplot objet
}
\description{
Function to get a histogram plot
}
