% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{get_voltage_histogram_plot_parameters}
\alias{get_voltage_histogram_plot_parameters}
\title{Function to get voltage histogram ggplot parameters, this may take environment parameters in future}
\usage{
get_voltage_histogram_plot_parameters()
}
\value{
list
}
\description{
Function to get voltage histogram ggplot parameters, this may take environment parameters in future
}
