% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_zone_substation}
\alias{get_zone_substation}
\title{Function to get all zone substations}
\usage{
get_zone_substation()
}
\value{
dataframe
}
\description{
Function to get all zone substations
}
\examples{
get_zone_substation()
}
