% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_zone_substation_channels}
\alias{get_zone_substation_channels}
\title{Function to get channels under a zone substation}
\usage{
get_zone_substation_channels(zone_substation)
}
\arguments{
\item{zone_substation}{The zone substation label}
}
\value{
dataframe
}
\description{
Function to get channels under a zone substation
}
\examples{
get_zone_substation_channels('CDA')
}
