% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_zone_substation_distribution_substation}
\alias{get_zone_substation_distribution_substation}
\title{Function to get distribution substations under a zone substation}
\usage{
get_zone_substation_distribution_substation(zone_substation)
}
\arguments{
\item{zone_substation}{The zone substation label}
}
\value{
Dataframe
}
\description{
Function to get distribution substations under a zone substation
}
\examples{
get_zone_substation_distribution_substation(zone_substation = 'CDA')
}
