% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{parse_device_string}
\alias{parse_device_string}
\title{Function to parse out device input into channels, meters, serialnums, nmis}
\usage{
parse_device_string(input_text)
}
\arguments{
\item{input_text}{A string containing channels, nic mac ids, serial numbers and nmis each on it's own line}
}
\value{
list containing channels,meters,serial_nums,nmis
}
\description{
Function to parse out device input into channels, meters, serialnums, nmis
}
\examples{
parse_device_string("0013500300101ce4_b\\n0013500300135a31_a\\n00135003001144df\\n01254662\\n6407275764")
parse_device_string("0013500300101ce4_b")
}
