% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{perform_get_request}
\alias{perform_get_request}
\title{Function to perform GET request with the supplied url}
\usage{
perform_get_request(request_url)
}
\arguments{
\item{request_url}{Url of get request}
}
\value{
Response object
}
\description{
Function to perform GET request with the supplied url
}
\examples{
perform_get_request(request_url = "http://rtssbx005.domain.dev.int:8889/hierarchy/meter/meter/0013500300086c70")
}
