% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{perform_post_request}
\alias{perform_post_request}
\title{Function to perform POST request with the supplied url, request body}
\usage{
perform_post_request(request_url, body)
}
\arguments{
\item{request_url}{Url of get request}

\item{body}{List, body of request}
}
\value{
Response object
}
\description{
Function to perform POST request with the supplied url, request body
}
\examples{
perform_post_request(request_url = "http://rtssbx005.domain.dev.int:8889/hierarchy/meter/meter", body = list('meters' = as.list("0013500300086c70")))
}
